feat(server): used transaction in user provider for sign up function and created association methods in user provider
