+++ apps/server/src/interfaces/role.interface.ts
@@ -2,3 +2,4 @@ export interface RoleInterface {
 	id?: number;
 	name: string;
 }
+export type RoleType = 'admin' | 'user';
