+++ apps/server/src/providers/user/create.ts
@@ -1,51 +1,30 @@
 import { UserProvider } from '.';
 import { User } from '../../models/user/index';
-import { roleProvider } from './roles';
+import { RoleType } from '../../interfaces/role.interface';
+import { sequelize } from '../../config/db/db';
 export const createUser = async ({
 	username,
 	email,
 	password,
-	isAdmin = false,
+	role = 'user',
 }: {
 	username: string;
 	email: string;
 	password: string;
-	isAdmin?: boolean;
+	role?: RoleType;
 }) => {
+	const transaction = await sequelize.transaction();
 	try {
-		const newUser = User.build({ username, email, password });
+		const user = User.build({ username, email, password });
 
-		await newUser.validate();
+		const userWithRole = await UserProvider.AssociateWith.role.add(user, role);
 
-		await newUser.save();
+		await userWithRole.save({ transaction });
 
-		let userRole = null;
-
-		if (!isAdmin) {
-			userRole = await roleProvider.get.byName('user');
-			if (!userRole) {
-				console.log(userRole);
-				throw new Error('User role not found');
-			}
-		} else {
-			userRole = await roleProvider.get.byName('admin');
-			if (!userRole) {
-				console.log(userRole);
-				throw new Error('User role not found');
-			}
-		}
-
-		if (userRole?.id) {
-			newUser.roleId = userRole.id;
-			await newUser.save();
-		}
-
-		const newUserFromDb = await UserProvider.getUser.ByEmail(email);
-		if (!newUserFromDb) {
-			return Promise.reject('User not found');
-		}
-		return Promise.resolve(newUserFromDb);
+		await transaction.commit();
+		return Promise.resolve(userWithRole);
 	} catch (error) {
+		await transaction.rollback();
 		return Promise.reject(error);
 	}
 };
