+++ apps/server/src/routes/roles/index.ts
@@ -1,4 +1,4 @@
-import { rolesControllers } from '../../controllers/user/roles';
+import { rolesControllers } from '../../controllers/roles';
 import express from 'express';
 // validators
 import { idValidator } from '../../middlewares/validators';
