+++ apps/server/src/services/database.ts
@@ -6,7 +6,7 @@ import { DietaryInterface } from '../interfaces/dietary.interface';
 
 import { readFile } from 'node:fs/promises';
 import path from 'node:path';
-import { DietaryProvider } from '../providers/user/dietary';
+import { DietaryProvider } from '../providers/dietary';
 
 async function syncDb() {
 	try {
@@ -51,7 +51,7 @@ async function seedDbUsers() {
 				username: 'culinaryalchemy',
 				email: 'culinaryalchemyofficial@gmail.com',
 				password: process.env.ADMIN_PASSWORD!,
-				isAdmin: true,
+				role: 'admin'
 			});
 		}
 		if (!doesOfficialUserExist && !officialUser) {
